+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support rendering on pixmaps larger than 8191x8191 pixels.
+  From now, `Pixmap` is limited only by the amount of memory caller has.
 
 ## [0.6.0] - 2021-08-21
 ### Added